+++ src/postgraphql/__tests__/postgraphqlIntegrationSchemaExport-test.js
@@ -1,6 +1,6 @@
 jest.mock('fs')
 
-import withPgClient from '../../postgres/__tests__/fixtures/withPgClient'
+import withPgClient from '../../../__tests__/utils/withPgClient'
 import createPostGraphQLSchema from '../schema/createPostGraphQLSchema'
 import exportPostGraphQLSchema from '../schema/exportPostGraphQLSchema'
 import { writeFile } from 'fs'
