+++ src/postgres/inventory/collection/__tests__/PgCollection-test.js
@@ -1,8 +1,8 @@
 import { Client } from 'pg'
 import { NullableType } from '../../../../interface'
-import withPgClient from '../../../__tests__/fixtures/withPgClient'
-import pgPool from '../../../__tests__/fixtures/pgPool'
-import kitchenSinkSchemaSql from '../../../__tests__/fixtures/kitchenSinkSchemaSql'
+import withPgClient from '../../../../__tests__/utils/withPgClient'
+import pgPool from '../../../../__tests__/utils/pgPool'
+import kitchenSinkSchemaSql from '../../../../__tests__/utils/kitchenSinkSchemaSql'
 import { mapToObject } from '../../../utils'
 import { PgCatalog, introspectDatabase } from '../../../introspection'
 import { $$pgClient } from '../../pgClientFromContext'
